requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.5.7
Faker==19.0.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
python-dateutil==2.8.2
selenium==4.10.0
six==1.16.0
urllib3==2.0.3
